range editing for new/existing preset
